feat: spm dependecy with swift and objc wrapper added

+++ composeApp/src/iosMain/kotlin/com/rbbozkurt/demo/random/NumberGenerator.ios.kt
@@ -1,14 +1,13 @@
 package com.rbbozkurt.demo.random
 
-import RNG.RandomNumberGenerator
+import RNGSwift.RandomNumberGenerator
 import kotlinx.cinterop.ExperimentalForeignApi
-
 @OptIn(ExperimentalForeignApi::class)
 class IosNumberGenerator : NumberGenerator {
     val randomNumberGenerator = RandomNumberGenerator()
     override fun generate(min: Int, max: Int): Int {
         return randomNumberGenerator
-            .generateWithMin(min.toLong(), max.toLong()).toInt()
+            .generateWithMin(min, max)
     }
 
 }
